backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow>=11.0.0
django-filter==23.5
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
django-storages>=1.14.2
django-cloudinary-storage>=0.3.0
cloudinary>=1.36.0